classing everythign to deploy adding queue for verify removing ts-node using tsx instead
